Incorporated code review comments.
