+++ machine-learning/solutions/kubeflow-pipelines-sentiment-analysis/README.md
@@ -2,15 +2,13 @@ Sentiment Analysis with Kubeflow Pipelines, Cloud Dataflow, and Cloud Natural La
 
 Included code will build a Kubeflow Pipelines component and pipeline. The pipeline uses Cloud Dataflow to do sentiment analysis on New York Times headlines. Cloud DAtaflow uses Apache Beam (Java) to extract front page headlines from NYTimes Archive API JSON files, perform sentiment analysis on them by calling Cloud Natural Language API and then aggregates the sentiment scores and magnitudes.
 
-# To create Sentiment Analysis docker image, open `./sentiment-analysis/build_image.sh` and set your Project ID. Then in the command line run
+# To create Sentiment Analysis docker image, go to `./sentiment-analysis` folder and run the following command replacing `PROJECT_ID` with your actual Project ID.
 
-    cd sentiment-analysis
-    bash build_image.sh
+    bash build_image.sh PROJECT_ID
 
-# To create Sentiment Analysis pipeline, in the command line run
+# To create Sentiment Analysis pipeline, go to `./pipeline` folder and run the following command.
 
-    cd pipeline
-    python build_sentiment_analysis.python
+    python build_sentiment_analysis.py
 
 # Other steps that must be completed
