+++ machine-learning/solutions/kubeflow-pipelines-sentiment-analysis/sentiment-analysis/Dockerfile
@@ -1,6 +1,5 @@
 FROM maven:3.6.0-jdk-8
 WORKDIR /beam
-COPY README.md /beam/
 COPY pom.xml /beam/
 COPY src /beam/src
 COPY Dockerfile /beam/
